refactor: test을 위한 모듈을 test로 분리
