+++ detector-core/src/main/java/io/jeyong/core/aspect/factory/interceptor/ConnectionCloseInterceptor.java
@@ -1,4 +1,4 @@
-package io.jeyong.core.interceptor;
+package io.jeyong.core.aspect.factory.interceptor;
 
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
