+++ detector-core/src/main/java/io/jeyong/core/query/context/QueryContext.java
@@ -1,4 +1,4 @@
-package io.jeyong.core.context;
+package io.jeyong.core.query.context;
 
 import java.util.HashMap;
 import java.util.Map;
