+++ detector-test/build.gradle
@@ -14,6 +14,6 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-aop:3.0.0'
-    implementation 'org.hibernate.orm:hibernate-core:6.1.5.Final'
+    implementation project(':detector-core')
+    implementation 'org.springframework.boot:spring-boot-starter-test:3.0.0'
 }
