+++ detector-test/src/main/java/io/jeyong/test/exception/NPlusOneQueryException.java
@@ -1,12 +1,8 @@
-package io.jeyong.core.exception;
+package io.jeyong.test.exception;
 
 public class NPlusOneQueryException extends RuntimeException {
 
     public NPlusOneQueryException(final String query, final Long count) {
         super(String.format("N+1 query detected: '%s' was executed %d times.", query, count));
-    }
-
-    public NPlusOneQueryException(final String message) {
-        super(message);
     }
 }
