+++ java-test/src/test/java/io/jeyong/test/environment/test/exclude/ExcludeQueryTest.java
@@ -3,7 +3,7 @@ package io.jeyong.test.environment.test.exclude;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
-import io.jeyong.core.annotation.NPlusOneTest;
+import io.jeyong.test.annotation.NPlusOneTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
