+++ java-test/src/test/java/io/jeyong/test/environment/test/mode/configuration/ExceptionModeConfigurationTest.java
@@ -2,11 +2,11 @@ package io.jeyong.test.environment.test.mode.configuration;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import io.jeyong.core.annotation.NPlusOneTest;
 import io.jeyong.core.config.NPlusOneDetectorProperties;
-import io.jeyong.core.context.ExceptionContext;
-import io.jeyong.core.template.NPlusOneQueryCollector;
 import io.jeyong.core.template.NPlusOneQueryTemplate;
+import io.jeyong.test.annotation.NPlusOneTest;
+import io.jeyong.test.exception.context.ExceptionContext;
+import io.jeyong.test.template.NPlusOneQueryCollector;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
