+++ java-test/src/test/java/io/jeyong/test/environment/test/mode/integration/LoggingModeIntegrationTest.java
@@ -3,7 +3,7 @@ package io.jeyong.test.environment.test.mode.integration;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
-import io.jeyong.core.annotation.NPlusOneTest;
+import io.jeyong.test.annotation.NPlusOneTest;
 import io.jeyong.test.case2.service.ProductService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
